hybridplot using vector of double and merger fixes
git-svn-id: https://root.cern.ch/svn/root/branches/dev/roostats@26345 27541ba8-7e3a-0410-8455-c3a389f83636

+++ roofit/roostats/inc/HybridPlot.h
@@ -64,9 +64,9 @@ namespace RooStats {
       /// Constructor
       HybridPlot(const char* name,
                  const char* title,
-                 std::vector<float> sb_values,
-                 std::vector<float> b_values,
-                 float m2lnQ_data,
+                 std::vector<double> sb_values,
+                 std::vector<double> b_values,
+                 double m2lnQ_data,
                  int n_bins,
                  bool verbosity=true);
